move common middelewares to different file and test protected urls
