+++ backend/server.js
@@ -1,8 +1,11 @@
 const express = require("express");
 const dotenv = require("dotenv").config();
 const cookieParser = require("cookie-parser");
-const authController = require("./controllers/authController");
 const mongoose = require("mongoose");
+const authController = require("./controllers/authController");
+const { verifyTokenAndSetUser, isLoggedIn } = require("./middlewares/authenticationMiddleware");
+const {errorHandlers, notFound} = require("./middlewares/commonMiddleware");
+const userRoles = require('./enums/userRoles')
 
 const app = express();
 const port = process.env.PORT || 5000;
@@ -10,28 +13,19 @@ const port = process.env.PORT || 5000;
 app.use(cookieParser());
 app.use(express.json());
 
+app.use(verifyTokenAndSetUser);
+
 app.get("/", (req, res, next) => {
     console.log("Cookies: ", req.cookies.token);
-    
+
     res.sendStatus(200);
 });
 
+app.get("/protected", isLoggedIn(userRoles.BACKOFFICEUSER), (req, res, next) => {
+    res.status(201).json(req.user);
+});
 
 app.use("/auth", authController);
-
-function notFound(req, res, next) {
-    res.status(404);
-    const error = new Error('Not Found - ' + req.originalUrl);
-    next(error);
-  }
-
-function errorHandlers(err, req, res, next) {
-    res.status(res.statusCode || 500);
-    res.json({
-        message: err.message,
-        stack: process.env.ERRORSTACK === "true" ? err.stack : "",
-    });
-}
 
 app.use(notFound);
 app.use(errorHandlers);
